feat: enforce settings on any TOML file
