+++ docs/plugins.rst
@@ -12,19 +12,6 @@ Below are the currently included plugins.
 
 .. auto-generated-from-here
 
-.. _pyprojecttomlplugin:
-
-pyproject.toml
---------------
-
-Enforce config on `pyproject.toml (PEP 518) <https://www.python.org/dev/peps/pep-0518/#file-format>`_.
-
-See also `the [tool.poetry] section of the pyproject.toml file
-<https://github.com/python-poetry/poetry/blob/master/docs/docs/pyproject.md>`_.
-
-Style example: :ref:`Python 3.8 version constraint <example-python-3-8>`.
-There are :ref:`many other examples here <examples>`.
-
 .. _precommitplugin:
 
 .pre-commit-config.yaml
@@ -76,3 +63,18 @@ To check if ``some.txt`` file contains the lines ``abc`` and ``def`` (in any ord
 
     [["some.txt".contains]]
     line = "def"
+
+.. _tomlplugin:
+
+TOML files
+----------
+
+Enforce config on TOML files.
+
+E.g.: `pyproject.toml (PEP 518) <https://www.python.org/dev/peps/pep-0518/#file-format>`_.
+
+See also `the [tool.poetry] section of the pyproject.toml file
+<https://github.com/python-poetry/poetry/blob/master/docs/docs/pyproject.md>`_.
+
+Style example: :ref:`Python 3.8 version constraint <example-python-3-8>`.
+There are :ref:`many other examples here <examples>`.
