--- docs/source/nitpick.plugins.pyproject_toml.rst
@@ -1,7 +1,0 @@
-nitpick.plugins.pyproject\_toml module
-======================================
-
-.. automodule:: nitpick.plugins.pyproject_toml
-   :members:
-   :undoc-members:
-   :show-inheritance:
