+++ docs/source/nitpick.plugins.toml.rst
@@ -1,0 +1,7 @@
+nitpick.plugins.toml module
+===========================
+
+.. automodule:: nitpick.plugins.toml
+   :members:
+   :undoc-members:
+   :show-inheritance:
